Update features example

+++ examples/features/tidigits_preprocessor.py
@@ -1,12 +1,16 @@
 from __future__ import print_function, division, absolute_import
 
+from odin.utils import ArgController
+args = ArgController(
+).add('--reset', 'force re-run the wave convertion, and wave dataset preprocessing', False
+).parse()
+
 import matplotlib
 matplotlib.use('Agg')
 
 import os
-os.environ['ODIN'] = "cpu,float32,tensorflow"
+os.environ['ODIN'] = "cpu,float32"
 import shutil
-import cPickle
 
 import numpy as np
 
@@ -18,7 +22,7 @@ from odin.utils import get_all_files, get_all_ext, exec_commands
 README = \
 """
 Original sample rate: 20,000 Hz
-Downsampled sample rate: 8,000 Hz
+Downsampled sample rate: 16,000 Hz
 
 Category       Symbol    Number    Age Range (years)
   Man            M        111           21 - 70
@@ -28,7 +32,7 @@ Category       Symbol    Number    Age Range (years)
 
 Eleven digits were used:  "zero", "one", "two", ... , "nine", and "oh".
 
-Each speaker:
+Digits for each speaker:
     22 isolated digits (two tokens of each of the eleven digits)
     11 two-digit sequences
     11 three-digit sequences
@@ -36,20 +40,24 @@ Each speaker:
     11 five-digit sequences
     11 seven-digit sequences
 
-Example:
-     /data/adults/train/man/fd/6z97za.wav
-
+Example of original data:
+     => "/data/adults/train/man/fd/6z97za.wav"
      training material, adult male, speaker code "fd",
      digit sequence "six zero nine seven zero", "a" is first production.
 
-Example:
-     /data/adults/test/woman/pf/1b.wav
-
+     => "/data/adults/test/woman/pf/1b.wav"
      test material, adult female, speaker code "pf",
      digit sequence "one", "b" is second production.
 
-Saved WAV file:
-    /audio/train_g_08_17_as_a_4291815.wav
+Saved WAV file format:
+    * [train|test]
+    * [m|w|b|g] (alias for man, women, boy, girl)
+    * [age]
+    * [dialectID]
+    * [speakerID]
+    * [production]
+    * [digit_sequence]
+    => "train_g_08_17_as_a_4291815.wav"
 
     train material, child "g"irl, age is "08", dialect group is "17",
     speaker code "as", "a" is first production,
@@ -62,18 +70,18 @@ Saved WAV file:
 # ===========================================================================
 inpath = "/mnt/sdb1/TIDIGITS"
 outpath = '/home/trung/data/tidigits'
-wavpath = os.path.join(inpath, "tiaudio")
-wav_ds = os.path.join(inpath, "tiwave")
+wav_path = os.path.join(inpath, "wave")
+wav_ds = os.path.join(inpath, "raw")
 infopath = os.path.join(inpath, 'data/children/doc/spkrinfo.txt')
-USE_DOWNSAMPLED_DATASET = True
 
 print('Input path:', inpath)
-print('Convert WAV to:', wavpath)
+print('Convert to WAV at:', wav_path)
 print('WAVE dataset:', wav_ds)
 
 exts = get_all_ext(inpath)
-audio_files = get_all_files(inpath, filter_func=lambda f: f[-4:] == '.wav' and
-    f.split('/')[-3] in ('girl', 'boy', 'man', 'woman'))
+audio_files = get_all_files(inpath,
+                filter_func=lambda f: f[-4:] == '.wav' and
+                            f.split('/')[-3] in ('girl', 'boy', 'man', 'woman'))
 # ID     Gender     Age     Dialect    Usage
 # ID - Unique 2-character speaker identifier
 # Gender - (M-man, W-woman, B-boy, G-girl)
@@ -83,8 +91,16 @@ audio_files = get_all_files(inpath, filter_func=lambda f: f[-4:] == '.wav' and
 info = np.genfromtxt(infopath, dtype=str, skip_header=12)
 info = {ID.lower(): (Gender.lower(), Age, Dialect, Usage)
         for ID, Gender, Age, Dialect, Usage in info}
-gender_map = {"man": "m", "woman": "w", "boy": "b", "girl": "g"}
-usage_map = {"TST": "test", "TRN": "train"}
+gender_map = {
+    "man": "m",
+    "woman": "w",
+    "boy": "b",
+    "girl": "g"
+}
+usage_map = {
+    "TST": "test",
+    "TRN": "train"
+}
 
 
 def get_name(path):
@@ -104,26 +120,27 @@ def get_name(path):
 # Convert all SPHERE to wav using sph2pipe
 # ===========================================================================
 # ====== convert all compress audio to .wav using sph2pipe ====== #
-if False:
-    if os.path.exists(wavpath):
-        print("[WARNING] Remove old WAV audio at:", wavpath)
-        shutil.rmtree(wavpath)
-    os.mkdir(wavpath)
-    cmds = ["sph2pipe %s %s -f rif" % (path, os.path.join(wavpath, get_name(path)))
+if args.reset and os.path.exists(wav_path):
+    shutil.rmtree(wav_path)
+if not os.path.exists(wav_path):
+    os.mkdir(wav_path)
+    cmds = ["sph2pipe %s %s -f rif" % (path, os.path.join(wav_path, get_name(path)))
             for path in audio_files]
     exec_commands(cmds, print_progress=True)
 # ====== create new segments list ====== #
-segments = get_all_files(wavpath, filter_func=lambda f: f[-4:] == '.wav')
+segments = get_all_files(wav_path, filter_func=lambda f: f[-4:] == '.wav')
 # remove .wav extension
 segments = [[os.path.basename(path).replace('.wav', ''), path, 0, -1, 0]
             for path in segments]
 # ====== store everything in one dataset ====== #
-if False:
-    if os.path.exists(wav_ds):
-        shutil.rmtree(wav_ds)
-    wave = F.WaveProcessor(segments, output_path=wav_ds, audio_ext='.wav',
-                          sr_new=8000, dtype='float16', datatype='memmap',
-                          ncache=0.2, ncpu=12)
+if args.reset and os.path.exists(wav_ds):
+    shutil.rmtree(wav_ds)
+if not os.path.exists(wav_ds):
+    wave = F.WaveProcessor(segments, output_path=wav_ds,
+                           sr=None, sr_info={}, sr_new=16000,
+                           audio_ext='.wav', pcm=False, remove_dc_offset=True,
+                           maxlen=None, dtype='float16', datatype='memmap',
+                           ncache=0.2, ncpu=8)
     wave.run()
     with open(os.path.join(wav_ds, 'README'), 'w') as f:
         f.write(README)
@@ -151,34 +168,28 @@ for backend in ('odin', 'sptk'):
             title=os.path.basename(f))
     plot_save('/tmp/tmp_%s.pdf' % backend, dpi=180, clear_all=True)
 # ====== processing ====== #
-if USE_DOWNSAMPLED_DATASET:
-    segments = ds
-for backend in ('odin', 'sptk'):
-    path_ = outpath + '_%s' % backend
-    if os.path.exists(path_):
-        shutil.rmtree(path_)
-    print("Output path:", path_)
-    acous = F.SpeechProcessor(segments, path_,
-                    sr=None, sr_new=None,
-                    win=0.02, hop=0.01, window='hann',
-                    nb_melfilters=40, nb_ceps=12,
-                    get_spec=True, get_qspec=True,
-                    get_phase=True, get_pitch=True, get_f0=True,
-                    get_vad=2, get_energy=True, get_delta=2,
-                    fmin=64, fmax=None,
-                    preemphasis=0.97 if backend == 'odin' else None,
-                    pitch_threshold=0.3, pitch_fmax=260,
-                    vad_smooth=12, vad_minlen=0.1,
-                    cqt_bins=96, pca=True, pca_whiten=False,
-                    save_stats=True, substitute_nan=None,
-                    power=2, log=True, backend=backend,
-                    audio_ext='.wav', dtype='float16', datatype='memmap',
-                    ncache=0.2, ncpu=12 + 1)
-    acous.run()
-    # copy README
-    with open(os.path.join(path_, 'README'), 'w') as f:
-        f.write(README)
-    # ====== validate saved dataset ====== #
-    ds = F.Dataset(path_, read_only=True)
-    print(ds.readme)
-    ds.close()
+acous = F.SpeechProcessor(ds, outpath,
+                sr=None, sr_info={}, sr_new=None,
+                win=0.03, hop=0.01, window='hann',
+                nb_melfilters=40, nb_ceps=13,
+                get_spec=True, get_qspec=False, get_phase=False,
+                get_pitch=True, get_f0=True,
+                get_vad=True, get_energy=True, get_delta=2,
+                fmin=64, fmax=None,
+                pitch_threshold=0.3, pitch_fmax=260, pitch_algo='swipe',
+                vad_smooth=3, vad_minlen=0.1,
+                cqt_bins=96, preemphasis=0.97,
+                center=True, power=2, log=True, backend='odin',
+                pca=True, pca_whiten=False,
+                audio_ext=None, maxlen=None, save_raw=True,
+                save_stats=True, substitute_nan=None,
+                dtype='float16', datatype='memmap',
+                ncache=120, ncpu=8)
+acous.run()
+# copy README
+with open(os.path.join(outpath, 'README'), 'w') as f:
+    f.write(README)
+# ====== validate saved dataset ====== #
+ds = F.Dataset(outpath, read_only=True)
+print(ds)
+ds.close()
